# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: IoC Cognition Fabric Node service — shared memory routing and memory
    operations proxy.
  title: CFN Service API
  version: "1.0"
paths:
...
  /api/workspaces/{workspaceId}/multi-agentic-systems/{masId}/shared-memories:
    post:
      consumes:
      - application/json
      description: Creates or updates shared memories with entries (concepts and relations)
        extracted from the provided trace or OpenClaw output for a given workspace
        and multi-agentic system.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Create or update shared memories request
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Shared memories successfully created or updated
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update shared memories.
      tags:
      - shared-memories
  /api/workspaces/{workspaceId}/multi-agentic-systems/{masId}/shared-memories/query:
    post:
      consumes:
      - application/json
      description: Queries shared memories for a given workspace and multi-agentic
        system using a graph path query.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Query request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Query executed successfully
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch shared memories
      tags:
      - shared-memories
swagger: "2.0"